+++ cluster-worker/pkg/reconcile/worker.go
@@ -6,6 +6,7 @@ import (
 	"context"
 	"fmt"
 	"log/slog"
+	"sync/atomic"
 	"time"
 
 	"github.com/fundament-oss/fundament/cluster-worker/pkg/handler"
@@ -21,6 +22,7 @@ type Worker struct {
 	registry *handler.Registry
 	logger   *slog.Logger
 	cfg      Config
+	ready    atomic.Bool
 }
 
 func New(registry *handler.Registry, logger *slog.Logger, cfg Config) *Worker {
@@ -31,12 +33,18 @@ func New(registry *handler.Registry, logger *slog.Logger, cfg Config) *Worker {
 	}
 }
 
+// IsReady returns true after the first reconciliation has completed.
+func (w *Worker) IsReady() bool {
+	return w.ready.Load()
+}
+
 // Run starts the reconciliation loop. It should be run as a separate goroutine.
 func (w *Worker) Run(ctx context.Context) error {
 	w.logger.Info("starting reconcile loop", "interval", w.cfg.Interval)
 
 	// Run immediately on startup, then on the ticker interval.
 	w.reconcileAll(ctx)
+	w.ready.Store(true)
 
 	ticker := time.NewTicker(w.cfg.Interval)
 	defer ticker.Stop()
